+++ bootmachine/contrib/distros/rackspace_arch_201208.py
@@ -1,13 +1,33 @@
 import time
+import urllib2
 
 from fabric.api import env, run, sudo
 from fabric.context_managers import cd, settings as fabric_settings
-from fabric.contrib.files import append, sed, uncomment
+from fabric.contrib.files import append, contains, sed, uncomment
 from fabric.operations import reboot
+from fabric.utils import abort
+
+import settings
 
 
 DISTRO = "ARCH_201208"
 SALT_INSTALLERS = ["aur", "aur-git"]
+
+
+def validate_configurator_version():
+    """
+    Arch is a rolling release distro, therefore it is important to ensure
+    the configurator version is current.
+    """
+    if settings.CONFIGURATOR_MODULE == "bootmachine.contrib.configurators.salt":
+        pkgver = settings.AUR_PKGVER
+        pkgrel = settings.AUR_PKGREL
+        response = urllib2.urlopen("https://aur.archlinux.org/packages/sa/salt/PKGBUILD")
+        for line in response:
+            if line.startswith("pkgver=") and not pkgver in line:
+                abort("The requested Salt 'pkgrel={0}' in the AUR was updated to '{1}'.".format(pkgver, line.strip()))
+            if line.startswith("pkgrel=") and not pkgrel in line:
+                abort("The requested Salt 'pkgrel={0}' in the AUR was updated to '{1}'.".format(pkgrel, line.strip()))
 
 
 def bootstrap():
@@ -16,8 +36,13 @@ def bootstrap():
 
     Only the bare essentials, the configurator will take care of the rest.
     """
-    # pre upgrade maintenance (updating filesystem and tzdata before pacman)
-    run("pacman --noconfirm -Syyu")
+    validate_configurator_version()
+    # configure kernel
+    sed("/etc/mkinitcpio.conf", "xen-", "xen_")  # see: https://projects.archlinux.org/mkinitcpio.git/commit/?id=5b99f78331f567cc1442460efc054b72c45306a6
+    sed("/etc/mkinitcpio.conf", "usbinput", "usbinput fsck")
+
+    # pre upgrade maintenance
+    run("pacman --noconfirm -Syu")
 
     # install essential packages
     run("pacman --noconfirm -S base-devel")
@@ -29,11 +54,11 @@ def bootstrap():
     # create a user, named 'aur', to safely install AUR packages under fakeroot
     # uid and gid values auto increment from 1000
     # to prevent conficts set the 'aur' user's gid and uid to 902
-    run("groupadd -g 902 aur && useradd -u 902 -g 902 -G wheel aur")
+    run("groupadd -g 902 aur && useradd -m -u 902 -g 902 -G wheel aur")
     uncomment("/etc/sudoers", "wheel.*NOPASSWD")
 
     # upgrade non-pacman rackspace installed packages
-    with cd("/tmp/"):
+    with cd("/home/aur/"):
         sudo("yaourt --noconfirm -S xe-guest-utilities", user="aur")
 
     # upgrade grub
@@ -44,24 +69,36 @@ def bootstrap():
     run("grub-install --directory=/usr/lib/grub/i386-pc --target=i386-pc --boot-directory=/boot --recheck --debug /dev/xvda")
     run("grub-mkconfig -o /boot/grub/grub.cfg")
 
-    # tweak sshd_config (before reboot so it is restarted!) so fabric can sftp with contrib.files.put, see:
+    # allow fabric to sftp with contrib.files.put
     # http://stackoverflow.com/questions/10221839/cant-use-fabric-put-is-there-any-server-configuration-needed
+    # change before reboot because then the sshd config will be reloaded
     sed("/etc/ssh/sshd_config", "Subsystem sftp /usr/lib/openssh/sftp-server", "Subsystem sftp internal-sftp")
 
-    # configure new kernel before reboot
-    sed("/etc/mkinitcpio.conf", "xen-", "xen_")  # see: https://projects.archlinux.org/mkinitcpio.git/commit/?id=5b99f78331f567cc1442460efc054b72c45306a6
-    sed("/etc/mkinitcpio.conf", "usbinput", "usbinput fsck")
-    run("mkinitcpio -p linux")
-
     # a pure systemd installation
+    run("printf 'y\ny\nY\n' | pacman -S systemd ntp")
+    sed("/etc/default/grub", 'GRUB_CMDLINE_LINUX=""', 'GRUB_CMDLINE_LINUX="init=/usr/lib/systemd/systemd"')
+    run("grub-mkconfig -o /boot/grub/grub.cfg")
+    run("iptables-save > /etc/iptables/iptables.rules")
+    run("ip6tables-save > /etc/iptables/ip6tables.rules")
+    for daemon in ["netcfg", "sshd", "syslog-ng", "ntpd", "iptables"]:
+        run("systemctl enable {0}.service".format(daemon))
     server = [s for s in env.bootmachine_servers if s.public_ip == env.host][0]
     append("/etc/hostname", server.name)
     append("/etc/locale.conf", "LANG=en_US.UTF-8\nLC_COLLATE=C")
-    run("printf 'y\nY\n' | pacman -S systemd-sysvcompat")
-    reboot()
-    run("pacman --noconfirm -Rns initscripts")
-    for daemon in ["netcfg", "sshd", "syslog-ng"]:
-        run("systemctl enable {0}.service".format(daemon))
+    with fabric_settings(warn_only=True):
+        reboot()
+    if not contains("/proc/1/comm", "systemd"):
+        abort("systemd installation failure")
+    run("timedatectl set-timezone US/Central")
+    sed("/etc/default/grub", 'GRUB_CMDLINE_LINUX="init=/usr/lib/systemd/systemd"', 'GRUB_CMDLINE_LINUX=""')
+    run("grub-mkconfig -o /boot/grub/grub.cfg")
+    run("pacman --noconfirm -Rns initscripts sysvinit")
+    run("pacman --noconfirm -S systemd-sysvcompat")
+
+    # ensure systemd, kernel, etc is truly up-to-date
+    run("pacman --noconfirm -Syu")
+    with fabric_settings(warn_only=True):
+        reboot()
 
 
 def install_salt(installer="aur"):
@@ -70,9 +107,9 @@ def install_salt(installer="aur"):
     """
     append("/etc/hosts",
            "{0} saltmaster-private".format(env.master_server.private_ip))
-
-    with cd("/tmp/"):
+    with cd("/home/aur/"):
         if installer == "aur":
+            validate_configurator_version()
             sudo("yaourt --noconfirm -S salt", user="aur")
         elif installer == "aur-git":
             sudo("yaourt --noconfirm -S salt-git", user="aur")
@@ -87,17 +124,20 @@ def setup_salt():
     """
     server = [s for s in env.bootmachine_servers if s.public_ip == env.host][0]
 
-    run("cp /etc/salt/minion.template /etc/salt/minion")
     if env.host == env.master_server.public_ip:
-        run("cp /etc/salt/master.template /etc/salt/master")
+        run("touch /etc/salt/master")
+        append("/etc/salt/master", "file_roots:\n  base:\n    - {0}".format(
+               settings.REMOTE_STATES_DIR))
+        append("/etc/salt/master", "pillar_roots:\n  base:\n    - {0}".format(
+               settings.REMOTE_PILLARS_DIR))
         run("systemctl enable salt-master.service")
-    run("systemctl enable salt-minion.service")
-
-    sed("/etc/salt/minion", "#master: salt", "master: {0}".format(env.master_server.private_ip))
-    sed("/etc/salt/minion", "#id:", "id: {0}".format(server.name))
+    run("touch /etc/salt/minion")
+    append("/etc/salt/minion", "master: {0}".format(env.master_server.private_ip))
+    append("/etc/salt/minion", "id: {0}".format(server.name))
     append("/etc/salt/minion", "grains:\n  roles:")
     for role in server.roles:
         append("/etc/salt/minion", "    - {0}".format(role))
+    run("systemctl enable salt-minion.service")
 
 
 def start_salt():