+++ bootmachine/settings_tests.py
@@ -5,15 +5,20 @@ import os
 CONFIGURATION MANAGEMENT
 """
 # salt
-SALTSTATES_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)),
-                              "configuration", "bootmachine-salt-states/")
-PILLAR_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)),
-                          "configuration", "bootmachine-pillar/")
+LOCAL_STATES_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)),
+                                "configuration", "salt", "states/")
+LOCAL_PILLARS_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)),
+                                 "configuration", "salt", "pillars/")
+REMOTE_STATES_DIR = "/srv/salt/states/"
+REMOTE_PILLARS_DIR = "/srv/salt/pillars/"
 SALT_INSTALLER_ARCH_201208 = "aur"
+AUR_PKGVER = "0.10.5"
+AUR_PKGREL = "4"
 SALT_INSTALLER_DEBIAN_6 = "backports"
 SALT_INSTALLER_FEDORA_16 = "rpm-stable"
 SALT_INSTALLER_FEDORA_17 = "rpm-stable"
 SALT_INSTALLER_UBUNTU_1204LTS = "ppa"
+SALT_INSTALLER_UBUNTU_1210 = "ppa"
 
 # puppet (not yet implemented)
 PUPPET_VERSION = NotImplementedError()
@@ -22,6 +27,7 @@ PUPPET_RECIPES_DIR = NotImplementedError()
 # chef (not yet implemented)
 CHEF_VERSION = NotImplementedError()
 CHEF_RECIPIES_DIR = NotImplementedError()
+
 
 """
 PROVIDERS AND SERVER STACK
@@ -41,6 +47,7 @@ OS_COMPUTE_API_VERSION = os.environ.get("OS_COMPUTE_API_VERSION")
 # Amazon authentication via boto
 EC2_ACCESS_ID = os.environ.get("AWS_ACCESS_KEY_ID")
 EC2_SECRET_KEY = os.environ.get("AWS_SECRET_ACCESS_KEY")
+
 
 """
 SECURITY