+++ tests/rackspacev2-salt-debian6/settings.py
@@ -9,18 +9,18 @@ MASTER = "rackspacev2-salt-debian6-a"
 PROVIDER_MODULE = "bootmachine.contrib.providers.rackspace_openstack_v2"
 SERVERS = [
     {"servername": "rackspacev2-salt-debian6-a",
-     "roles": ["loadbalancer"],
+     "roles": ["salt-master"],
      "flavor": "2",
      "image": "Debian 6 (Squeeze)",
      "distro_module": "bootmachine.contrib.distros.rackspace_debian_6"},
     {"servername": "rackspacev2-salt-debian6-b",
-     "roles": ["application"],
+     "roles": ["test-additional-minions"],
      "flavor": "2",
      "image": "Debian 6 (Squeeze)",
      "distro_module": "bootmachine.contrib.distros.rackspace_debian_6"},
 ]
 
-SSH_PUBLIC_KEY = os.path.join(os.environ["HOME"], ".ssh", "id_ecdsa.pub")
+SSH_PUBLIC_KEY = os.path.join(os.environ["HOME"], ".ssh", "id_rsa.pub")
 enc, keyhash = open(SSH_PUBLIC_KEY).read().split(" ")[:2]
 if "==" in keyhash:
     keyhash = keyhash.split("==")[0]